test: relocate non-asserting tests under Tests/IntegrationTests/<module>/
Tests in Tests/IntegrationTests/ are exploratory dumps without assertions;
keeping them under per-module subfolders (MachOSwiftSection, MachOSymbols,
SwiftDump, SwiftInspection, SwiftInterface, TypeIndexing) makes the split
between unit tests and manual-inspection tests obvious. Drop a few
demangling/symbol dumps superseded by fixture-based coverage, and move the
real LayoutTests unit test into Tests/MachOSwiftSectionTests/ where it
belongs. CLAUDE.md records the convention so agents skip IntegrationTests
in --skip runs.
