+++ Tests/IntegrationTests/MachOSwiftSection/PrimitiveTypeMappingTests.swift
