+++ Tests/IntegrationTests/MachOSwiftSection/ProtocolRequirementSignatureTests.swift
