+++ Tests/IntegrationTests/SwiftInspection/ClassHierarchyDumpTests.swift
