+++ Tests/IntegrationTests/SwiftInspection/MultiPayloadEnumTests.swift
