+++ Tests/IntegrationTests/TypeIndexing/SourceKitManagerTests.swift
