--- Tests/MachOSymbolsTests/DemanglingTests.swift
@@ -1,132 +1,0 @@
-import Foundation
-import Testing
-@testable import Demangling
-import MachOKit
-import MachOFoundation
-@testable import MachOTestingSupport
-import MachOFixtureSupport
-import Dependencies
-
-@MainActor
-protocol DemanglingTests {
-    func allSymbols() async throws -> [MachOSwiftSymbol]
-    func mainTest() async throws
-}
-
-extension DemanglingTests {
-    func mainTest() async throws {
-        let allSwiftSymbols = try await allSymbols()
-
-        // Counters
-        var totalCount = 0
-        var successCount = 0
-        var knownIssueCount = 0
-        var demangleFailCount = 0
-        var nodeTreeMismatchCount = 0
-        var remangleMismatchCount = 0
-
-        // Sample collection (limit per category)
-        let maxSamples = 10
-        var demangleFailSamples: [String] = []
-        var nodeTreeMismatchSamples: [String] = []
-        var remangleMismatchSamples: [String] = []
-
-        for symbol in allSwiftSymbols {
-            totalCount += 1
-            let mangledName = symbol.stringValue
-            let stdlibTree = MachOFixtureSupport.stdlib_demangleNodeTree(mangledName)
-
-            do {
-                let node = try await demangleAsNode(mangledName)
-                var allPassed = true
-
-                // 1. Node tree check
-                if let stdlibTree {
-                    let ourTree = node.description + "\n"
-                    if stdlibTree != ourTree {
-                        if isOpaqueReturnTypeParentDifference(stdlibTree, ourTree) {
-                            knownIssueCount += 1
-                        } else {
-                            allPassed = false
-                            nodeTreeMismatchCount += 1
-                            if nodeTreeMismatchSamples.count < maxSamples {
-                                nodeTreeMismatchSamples.append(mangledName)
-                            }
-                            Issue.record("Node tree mismatch: \(mangledName)")
-                        }
-                    }
-                }
-
-                // 2. Remangle check
-                let remangled = try await Demangling.mangleAsString(node)
-                if remangled != mangledName {
-                    // Known issue: Md vs MD (Apple-internal lowercase 'd')
-                    if mangledName.hasSuffix("Md"), remangled.hasSuffix("MD"),
-                       mangledName.dropLast(2) == remangled.dropLast(2) {
-                        knownIssueCount += 1
-                    } else {
-                        allPassed = false
-                        remangleMismatchCount += 1
-                        if remangleMismatchSamples.count < maxSamples {
-                            remangleMismatchSamples.append("  \(mangledName)\n    remangled: \(remangled)")
-                        }
-                        Issue.record("Remangle mismatch: \(mangledName)")
-                    }
-                }
-
-                if allPassed { successCount += 1 }
-            } catch {
-                if stdlibTree != nil {
-                    // stdlib succeeded but we failed
-                    demangleFailCount += 1
-                    if demangleFailSamples.count < maxSamples {
-                        demangleFailSamples.append("  \(mangledName) — \(error)")
-                    }
-                    Issue.record("Demangle failed: \(mangledName): \(error)")
-                } else {
-                    successCount += 1 // both failed = consistent
-                }
-            }
-        }
-
-        // Print summary
-        print("""
-
-        ═══ Demangling Alignment Report ═══
-        Total symbols:         \(totalCount)
-        Passed:                \(successCount)
-        Known issues (skip):   \(knownIssueCount)
-        Demangle failures:     \(demangleFailCount)
-        Node tree mismatches:  \(nodeTreeMismatchCount)
-        Remangle mismatches:   \(remangleMismatchCount)
-        """)
-
-        if !demangleFailSamples.isEmpty {
-            print("--- Demangle Failures (first \(demangleFailSamples.count)) ---")
-            for sample in demangleFailSamples {
-                print(sample)
-            }
-        }
-        if !nodeTreeMismatchSamples.isEmpty {
-            print("--- Node Tree Mismatches (first \(nodeTreeMismatchSamples.count)) ---")
-            for sample in nodeTreeMismatchSamples {
-                print(sample)
-            }
-        }
-        if !remangleMismatchSamples.isEmpty {
-            print("--- Remangle Mismatches (first \(remangleMismatchSamples.count)) ---")
-            for sample in remangleMismatchSamples {
-                print(sample)
-            }
-        }
-    }
-
-    /// Check if the difference between two tree strings is only in OpaqueReturnTypeParent lines.
-    private func isOpaqueReturnTypeParentDifference(_ lhs: String, _ rhs: String) -> Bool {
-        let filteredLhs = lhs.split(separator: "\n", omittingEmptySubsequences: false)
-            .filter { !$0.contains("OpaqueReturnTypeParent") }
-        let filteredRhs = rhs.split(separator: "\n", omittingEmptySubsequences: false)
-            .filter { !$0.contains("OpaqueReturnTypeParent") }
-        return filteredLhs == filteredRhs
-    }
-}
